requirements.txt
pymongo>=3.12.0
requests
python-dotenv